generate max-clients limit cleared event
Generate an event when max-clients limit condition is cleared.
This signals that another client can connect as there is an
available slot before max-clients limit is enforced again.

Because this is not a session specific event vsf_event_*
infrastructure was refactored to use an eventlog_fd instead
of grabbing it from the session structure.

Also, because vsf_event_max_clients_cleared() may be called
from signal context it builds the event message on stack to
avoid memory allocations before calling write().

vsf_event_log_err() function was removed as it requires access
to the session structure (that does not exist in case of
max-clients limit cleared) so if anything goes wrong with
sending out the event the error is no longer reported in the
logs.